FVM specific EVM precompiles - changes randomness functions to take a raw i64 since this is userspace now - add precompiles: resolve_address, lookup_address, get_actor_code_cid, get_randomness(beacon/chain)

+++ actors/miner/tests/miner_actor_test_wpost.rs
@@ -554,7 +554,7 @@ fn duplicate_proof_rejected() {
     // From version 7, a duplicate is explicitly rejected.
     rt.expect_validate_caller_addr(h.caller_addrs());
     rt.expect_get_randomness_from_tickets(
-        DomainSeparationTag::PoStChainCommit,
+        DomainSeparationTag::PoStChainCommit as i64,
         dlinfo.challenge,
         Vec::new(),
         commit_rand,
